libs, net: Introduce vmspec module and move helpers from tests to it
In order to keep the test files clean, easy to read and focused, it is
preferred to use library helpers.

The UDN tests have used some operations which are common to network
testing: Looking up for an iface in the VMI status.

In order to keep the test files clean, without too much noise and to
encourage usage of well defined operations on the VM manifest, a new
`vmspec` module has been created.
