+++ tests/network/user_defined_network/test_user_defined_network.py
@@ -4,8 +4,8 @@ import pytest
 from ocp_resources.resource import Resource
 from ocp_resources.user_defined_network import Layer2UserDefinedNetwork
 from ocp_resources.utils.constants import TIMEOUT_1MINUTE
-from timeout_sampler import TimeoutExpiredError, TimeoutSampler
 
+from libs.net.vmspec import VMInterfaceNotFoundError, get_iface
 from libs.vm import affinity
 from libs.vm.affinity import new_pod_anti_affinity
 from libs.vm.factory import base_vmspec, fedora_vm
@@ -39,68 +39,12 @@ def udn_primary_network(name):
     return Interface(name=name, binding=NetBinding(name="l2bridge")), Network(name=name, pod={})
 
 
-class VMInterfaceNotFoundError(Exception):
-    pass
-
-
 def vm_primary_network_name(vm):
     vm_primary_network_type = "pod"
     for network in vm.instance.spec.template.spec.networks:
         if vm_primary_network_type in network.keys():
             return network.name
     raise VMInterfaceNotFoundError(f"No interface connected to the primary network was found in VM {vm.name}.")
-
-
-def iface_lookup(vm, iface_name, predicate):
-    """
-    Returns the interface requested if found and the predicate function (to which the interface is
-    sent) returns True. Else, raise VMInterfaceNotFound.
-
-    Args:
-        vm (BaseVirtualMachine): VM in which to search for the network interface.
-        iface_name (str): The name of the requested interface.
-        predicate (function): A function that takes a network interface as an argument
-            and returns a boolean value. This function should define the condition that
-            the interface needs to meet.
-
-    Returns:
-        iface (dict): The requested interface.
-
-    Raises:
-        VMInterfaceNotFound: If the requested interface was not found in the VM.
-    """
-    for iface in vm.vmi.interfaces:
-        if iface.name == iface_name and predicate(iface):
-            return iface
-    raise VMInterfaceNotFoundError(f"Network interface named {iface_name} was not found in VM {vm.name}.")
-
-
-def wait_for_vm_iface(vm, iface_name, timeout, sleep, predicate=lambda iface: True):
-    samples = TimeoutSampler(
-        wait_timeout=timeout,
-        sleep=sleep,
-        func=iface_lookup,
-        vm=vm,
-        iface_name=iface_name,
-        predicate=predicate,
-        exceptions_dict={VMInterfaceNotFoundError: []},
-    )
-    for sample in samples:
-        if sample:
-            return sample
-
-
-def get_iface(vm, iface_name):
-    try:
-        return wait_for_vm_iface(
-            vm=vm,
-            iface_name=iface_name,
-            timeout=30,
-            sleep=5,
-            predicate=lambda interface: "guest-agent" in interface["infoSource"] and interface[IP_ADDRESS],
-        )
-    except TimeoutExpiredError:
-        raise VMInterfaceNotFoundError(f"Network interface named {iface_name} was not found in VM {vm.name}.")
 
 
 @pytest.fixture(scope="module")
